feat: add meta data to head, add editorsconfig, update folder structure
