+++ .editorconfig
@@ -1,0 +1,24 @@
+# EditorConfig is awesome: https://EditorConfig.org
+
+# Top-most EditorConfig file
+root = true
+
+[*]
+charset = utf-8
+end_of_line = lf
+indent_size = 4
+indent_style = space
+insert_final_newline = true
+trim_trailing_whitespace = true
+
+[*.css]
+indent_size = 4
+
+[*.scss]
+indent_size = 4
+
+[*.js]
+indent_size = 4
+
+[*.json]
+indent_size = 2
