+++ index.html
@@ -1,13 +1,16 @@
 <!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/png" href="/tree.png" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Green Ecolution</title>
-  </head>
-  <body>
-    <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
-  </body>
+<html lang="de">
+    <head>
+        <meta charset="UTF-8" />
+        <link rel="icon" type="image/png" href="/tree.png" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <title>Green Ecolution | Smartes Grünflächenmanagement</title>
+        <meta name="author" content="Hochschule Flensburg – Forschungsprojekt">
+        <meta name="description" content="Die Projektwebsite zum Forschungsprojekt »Green Ecolution« für smartes Grünflächenmanagement der HS Flensburg.">
+        <meta name="theme-color" content="#659858">
+    </head>
+    <body class="bg-white antialiased font-nunito-sans text-base text-grey-900 w-screen overflow-x-hidden lg:text-lg selection:bg-green-light-900/20 selection:text-grey-900">
+        <div id="root"></div>
+        <script type="module" src="/src/main.tsx"></script>
+    </body>
 </html>
